Generated output json filename includes timestamp and model name and goes in taxonomy dir (#51)
* Generated output json filename includes timestamp and model name and goes in taxonomy dir

Unless output_dir is explicitly passed in (but it is not currently a CLI argument anyway)...

The output generated data file will go in the taxonomy dir assuming it is a dir.
This commit also allows --taxonomy to refer to a file.  In that case, the output would go
in the same dir as that file. This is also the case if seed_file is used instead of taxonomy.

The output file is also timestamped and includes te model name (minus the path and file ext, if any).

The file is prefixed with "generated", has .json extension, and has underscore separators.

I.e.: generated_<model>_<isodatetime>.json
E.g.: generated_ggml-malachite-7b-Q4_K_M_2024-02-25T23:28:55.json

Also fixed the check for '.yaml' to add warning for '.yml' or upper case.

Fixes #38
Fixes #45

Signed-off-by: markstur <mark.sturdevant@ibm.com>

* REAMDE update for dated generate output (and optionally using regen.json as input)

Signed-off-by: markstur <mark.sturdevant@ibm.com>

---------

Signed-off-by: markstur <mark.sturdevant@ibm.com>