Generate unicode data without \u code points
Before the patch, training and test data that included unicode
characters was dumped with \u code points. This makes it a lot harder to
judge whether samples that include non-English characters are valid /
acceptable.

JSONL format requires that all strings are utf-8, so there should be no
ambiguity as to how to interpret the multibyte characters. (Note: I
believe JSONL still leaves space for escaped unicode characters even if
discourages using them, so what cli already does is not, strictly
speaking, a violation of the format.)

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>