lab: exit program after catching and reporting exceptions (#336)
* cli: use non-zero exit status for error scenarios

Invoking sys.exit() with no args results in an exit status
of zero, which traditionally indicates success. This is not
appropriate for error scenarios.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>

* lab: exit program after catching and reporting exceptions

Creating the app failed with an exception raised due to a missing
model file. Execution carried on, however, leading to use of an
undefined 'app' variable. Fix this, and a couple of other places
which also caught (unrecoverable) exceptions and forgot to exit.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>

---------

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>