Allow keyboard interrupt while getting and parsing response (#357)
Allow user to CTRL+C while getting a response.  Sometimes the API is
non-responsive or the response is taking a long time.  Allow the request
to be canceled without exiting chat.

This can be seen when the API is not serving content correctly and user
tries to cancel a request.  Or if you quickly type CTRL+C after typing
in a prompt.  An example of the behavior before this change:

```
>>> sing a song
^CUnknown error
Executing chat failed with: API issue found while executing chat:
Unknown error: <class 'KeyboardInterrupt'>
```
And the client exits.

After the change in this PR:

```
>>> sing a song
^C>>>
```
And the client is able to take additional prompts.

Note this does not have an impact of cancel while the response output is
being created.  Canceling that already works.

Signed-off-by: Naveen Malik <nmalik@redhat.com>