ci: Add flake8 for additional linting
Add a new tox environment for running `flake8` on the code. It can
enforce common Python formatting standards, but also catch many real
errors, as well.

To start, the command is configured to ignore every rule that is
currently hit. Each one disabled is documented in tox.ini. Future PRs
can clean some of them up.

When future changes trigger other rules, they will have to be fixed or
added to the ignore list, depending on the issue.

The new environment is also added to the linting github action.

Signed-off-by: Russell Bryant <rbryant@redhat.com>