ci: add newline linter
In review of #426, a reviewer caught that a file was missing a newline
at the end. This will save humans from having to point this out in the
future.

For now it's running only in existing workflows so it doesn't eat
additional github actions minutes. That means if a commit only changes
files not checked by these workflows (.github/* and */*.py), they will
get missed.

Signed-off-by: Russell Bryant <rbryant@redhat.com>