RFC: support dynamic taxonomies
For some taxonomies, it makes more sense to generate them
programmatically, which allows to generate a larger number of samples as
well as control their correctness without relying on humans.

For example, generating samples of simple rule application to an input,
that can be expressed as a simple pure function, e.g. transliteration
rules, is more robust and correct when taxonomies are not human written.
Some other taxonomies that lean more into knowledge territory may also
benefit from this.

This patch implements dynamic taxonomies as Containerfile definitions
that can generate the seed on invocation.

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>