Remove unused prompt.txt file from cli/generator/
The file with a custom prompt is read from the current directory when
calling the lab tool. I don't think it's common that the tool is called
from under the subdir of the source code, and regardless, the file is
supposed to be user generated (customized), not in-tree.

Adding the file name to .gitignore so that it's not committed later by
mistake.

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>