WIP: Communicate with the model using JSON
The text based communication with the model can be
inconsistent, resulting in much of the results being
dropped. See https://github.com/instruct-lab/cli/issues/466

This is an attempt to see if instructing it to generate
JSON is more consistent. Anecdotal evidence on small
instructions suggest there are less format misses then
the current regex.split approach. But this might not
be as reliable in instructions with more tokens as we
might be more likely to get in parsable JSON. More testing
needed here.

Signed-off-by: Derek Higgins <derekh@redhat.com>