Allow users to specify the number of CPU threads on lab serve (#423)
The underlying llama-cpp-python.server tooling defaults to using
the number of CPUs on the system divided by 2.  For GPU accelerated
inferencing, this is more than sufficient.  However, machines
that lack a capable GPU and must rely on only the CPU are limited
by this default.  This adds a --nr-threads option that allows the
user to specify how many CPU threads llama.cpp should use when
starting the server.

In local testing, this reduced the time for a default lab generate
run on a 16 thread AMD Ryzen CPU with 64GB of RAM from 7:51:33 to
6:48:46 with --nr-threads 14.  Any small bit helps.

Signed-off-by: Josh Boyer <jwboyer@redhat.com>