chat: Catch when context gets too long
We maintain chat history and pass it as context to the API. When that
context grew too long, chat would fail out saying an unknown error had
occurred and that it got a BadRequest exception.

The output from `lab serve` shows that it got a request with
context that was too long.

To resolve this, we just catch this specific error and trim the oldest
message from our history and try again. If we have nothing left to
trim and still get the error, we'll print an error to the console.

Closes #528

Signed-off-by: Russell Bryant <rbryant@redhat.com>