Fix "No such file ... ibm rebooks" during lab generate
PR https://github.com/instruct-lab/taxonomy/pull/259 renamed some files.
This is causing a regression in `lab check` and `lab generate`:

```
[Errno 2] No such file or directory: 'taxonomy/knowledge/technical
manual/ibm rebooks/qna.yaml'
```

The problem is caused by a bug in `get_taxonomy_diff()` function. It is
using `a_path` instead of `b_path`. The `a_path` is the original, old
name of a file and `b_path` is the new name. The new name can be `None`
when a file is deleted.

`inspect.getmembers()` of a renamed file:
```
  ...
  ('a_blob', <git.Blob "804ddc5713e1433dc0dd3f64036d0ba0215a24b3">),
  ('a_mode', 33188),
  ('a_path', 'knowledge/technical manual/ibm rebooks/qna.yaml'),
  ('a_rawpath', b'knowledge/technical manual/ibm rebooks/qna.yaml'),
  ('b_blob', <git.Blob "696aeb0b0c1db2be00aef545354f5f8b9c8dab87">),
  ('b_mode', 33188),
  ('b_path', 'knowledge/technical_manual/ibm_redbooks/qna.yaml'),
  ('b_rawpath', b'knowledge/technical_manual/ibm_redbooks/qna.yaml'),
  ('change_type', 'R'),
  ('copied_file', False),
  ('deleted_file', False),
  ('diff', ''),
  ('new_file', False),
  ('raw_rename_from', b'knowledge/technical manual/ibm rebooks/qna.yaml'),
  ('raw_rename_to', b'knowledge/technical_manual/ibm_redbooks/qna.yaml'),
  ('rename_from', 'knowledge/technical manual/ibm rebooks/qna.yaml'),
  ('rename_to', 'knowledge/technical_manual/ibm_redbooks/qna.yaml'),
  ('renamed', True),
  ('renamed_file', True),
  ('score', 98)]]
```

Signed-off-by: Christian Heimes <cheimes@redhat.com>