Make project reuse compliant
Add SPDX annotations for all files
in the project to make it reuse-compliant.

For those files that are not text-based (e.g., *png)
or cannot accept comments (e.g., *.json), a
filename.license file is created to ensure that
all resources are properly licensed.

```
$ reuse lint
\# SUMMARY

* Bad licenses: 0
* Deprecated licenses: 0
* Licenses without file extension: 0
* Missing licenses: 0
* Unused licenses: 0
* Used licenses: Apache-2.0
* Read errors: 0
* files with copyright information: 102 / 102
* files with license information: 102 / 102

Congratulations! Your project is compliant with version 3.0 of the REUSE
Specification :-)
```

Signed-off-by: Albert Esteve <aesteve@redhat.com>