Replace Black with Ruff
Ruff is an extremely fast Python linter and code formatter, written in
Rust. It is a fantastic tool that combines the features of flake8 linter,
isort, Black code formatter, and (to some degree) pylint in one tool.
It comes with over 800 linting rules. Ruff is not only blazing fast,
it can also fix most code violation. The code fixer is a fantastic quality
of life feature for developers.

Gradually introduce Ruff for linting and code formatting. Replaces Black
as code formatting tool. Ruff is configured to create the same results as
Black's default configuration.

Linters will be enabled in future commits.

See: #739
Co-authored-by: Miroslav Vadkerti <mvadkert@redhat.com>
Signed-off-by: Christian Heimes <cheimes@redhat.com>