fix: reduce client timeout while attempting to connect to server
When running `lab chat`, if the 8000 port is taken by another
application, the client's timeout to connect to the server is
set to 10min with a second retry. So the client could wait up to 20min
before we fallback to starting a temporary server.
This change sets a timeout to 30s and the retry of 2 remains unchanged.

Signed-off-by: Sébastien Han <seb@redhat.com>