fix: properly propagate errors when temp server fails to start
As reported in https://github.com/instruct-lab/cli/issues/683, the temp
server will fail to start if the 'lab chat' is called and the server is
not running. In this scenario, the temporary server is launched but
multiple exceptions could occur and they need to be reported cleanly.
One of them is the model being absent.

Fixes: https://github.com/instruct-lab/cli/issues/683
Signed-off-by: Sébastien Han <seb@redhat.com>