Allow lab test to run without a training run
I wanted a way to quickly generate sample output for the set of new
questions before model training. `lab test` was pretty close to what I
wanted. This adjusts the behavior so that if the adapter file is not
found, it will still run the test of model behavior from BEFORE
training.

Even if the adapter file is found, you can force it to not be used by
setting `--adapter-file none`.

Signed-off-by: Russell Bryant <rbryant@redhat.com>