fix: run the temp server with an available port (#744)
fix: run the temp server with an available port

If the server is not reachable and we fallback to running a temporary
server we need to pick a port that is available. Prior to this, we were
simply trying our luck by randomly picking a port within the (1024,
65535) range. In case we get unlucky and the port we picked already runs
a service, this would leave the chat with a binding error and a prompt
making it very confusing for the end user (since the server actually
never started).
Now, prior to running the temp server, we make sure the port is
available (at least at the time of the socket test) and use it to run
the server. If not, we try again and keep a list of tried ports, also we
fail in the unlikely event where no ports are available at all and we
tried all possible ones.

Signed-off-by: Sébastien Han <seb@redhat.com>