Restrict server to one client
The server frequently crashes when dealing with 2 clients simultaniously,
limit it to a single worker, If a user tries to connect ilab to the server
the second command will start its own local server (as it gets a http 503).

We also need to disable keep-alive to prevent consecutive calls in the
same client holding the only worker open and DOS'ing itself.

There is still one error case, if the chat client is started but not activly
chatting when generate is started. chat will fail when the model is called.
Not ideal but not as bad as an hour long "ilab generate" failing.

Fixes #346

Signed-off-by: Derek Higgins <derekh@redhat.com>