Add new `sdg-svc` subcommand
There is a service running on a private endpoint set up for
`@instruct-lab-bot` to use for generating sample synthetic data using
the full backend pipeline process. It makes sense to have the client
side of this in `icli`, because someday I could envision our generic
reusable pipeline providing this same type of service, and `icli`
would make sense as the client tool for it.

For now the help text does say this is experimental and an unstable
API. The command is also hidden so it does not show up in the command
list in the top-level help text.

Signed-off-by: Russell Bryant <rbryant@redhat.com>