Use the taxonomy schema to validate taxonomy yaml files (#776)
* fixes: Fix some minor errors found during development of schema support

Co-authored-by: Rafael Vasquez <rafvasq21@gmail.com>
Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>

* schema: Vendor in taxonomy schema files

We will later replace with a git submodule of the schema repo once
anonymous access to the repo is available. But for now, we vendor in
the files.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>

* schema: Use taxonomy schema to validate YAML files

We use jsonschema to validate that input yaml files conform to the
schema for their relative path. Any validation failures are logged
and the process ends with an error.

See https://github.com/instruct-lab/cli/issues/760

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>

---------

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>
Co-authored-by: Rafael Vasquez <rafvasq21@gmail.com>