fix: ignore SIGINT when running the temp server
When we simply run `ilab chat`, a temporary server will run in the
background and we can enjoy a great chat. However, if the chat
response is interrupted with ctrl+C, the signal will be caught by
uvicorn and then the server will be killed making further chat
impossible and printing a large backtrace.

Another option was to run the temp server in a dedicated thread but this
requires a bit more changes as well as the difficulty of handling the
thread termination. This would have been handled by uvicorn natively
since it is not setting up any signal handler but would still require
us to handle it manually in our code.

Signed-off-by: Sébastien Han <seb@redhat.com>