(fix) ilab diff for knowledge doc bug
Closes #885

Fix: `working_dir` arg being passed to `os.path.join()` for
`glob.glob` was of type `Repo`. Fix is to pass in `Repo.working_dir`,
which is of type `str` instead.

Also fixes an issue with `skip_checkout` variable, which was always
`False`. Fix is to use a temp variable `sc` and set `sc` to
`True/False` based on `skip_checkout`

Signed-off-by: Anik Bhattacharjee <anbhatta@redhat.com>