Fix unit test import mocking for python3.9
Mocks were not effective for in-function local import symbols.

This patch moves the imports to the module scope, which fixes the suite
for python3.9.

Since mlx is not present on Linux, we have to catch ImportError and fall
back to Nones for mac-specific imports (which can then be mocked out in
unit tests as needed).

Closes #814
Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>