Fix TypeError during `ilab generate`
Use the `working_dir` of a repo object, not the repo object as base
directory for globbing. Fixes:

```
  File "cli/utils.py", line 229, in get_documents
    for file_path in glob.glob(os.path.join(working_dir, pattern)):
                               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "<frozen posixpath>", line 76, in join
TypeError: expected str, bytes or os.PathLike object, not Repo
```

Replace the mock.patch with a proper spec-ed Mock that walks and talks
like a `git.Repo` object. This makes the code easier to understand and
does not violate the type annotations.

Fixes: #885
Signed-off-by: Christian Heimes <cheimes@redhat.com>