tests: fix input data file contents
Test code was writing {} in the file but this results in an empty dict()
produced by json.load, which would later fail with KeyError('user') in
make_data.

The reason why this doesn't show up in the test runs is because
make_data is mocked out. This test bug was noticed while working on
another fix for the test suite where mocks were not effective [1].

While the test suite was not failing (because of mocks), it's still
better to inject data that conforms to loader expectations.

[1] https://github.com/instructlab/instructlab/pull/890

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>