Refactor the CUDA container spec
This change modifies the Containerfile for NVIDIA CUDA to:
- use the CUDA runtime image for the final stage
- install the Python packages in a virtualenv that is copied in the
  final container
- make the final container rootless

It also proposes a shell alias that should simplify the user experience
when using the container image, avoiding to use long `podman run`
commands.

It modifies the containerization documentation to explain all of this
and provide steps for the user.

The `cuda` target in the Makefile is modified accordingly.

Signed-off-by: Fabien Dupont <fdupont@redhat.com>