Add mergify for PR automation
This PR introduces configuration for
[Mergify](https://docs.mergify.com), with an initial configuration
that will automatically merge PRs once they meet a set of defined
criteria.

What I like about this proposed configuration is that it allows the
repo maintainers to manage the explicit merge rules in a file in the
repo itself and then allow a tool to enforce them. The rules
configured in this PR would be:

- At least one mainatiner has approved the PR
- There are no outstanding review requests
- There are no requested changes outstanding
- A `hold` label is not on the PR
- The DCO check passes
- Other CI jobs (actionlint, lint, test) are required based on which
  files were changed in the PR.

Ideally once this configuration is in place, it should be incredibly
rare that anyone hits the "merge" button manually.

There are several other things that can be automated with Mergify -
check out the docs for more if you're interested.

Signed-off-by: Russell Bryant <rbryant@redhat.com>