fix four_bit_quant VRAM check for cuda
When using cuda and performing the VRAM check, the value was being
passed was either 11 or 17 and was meant to be an expession of
gibibytes. However, it was never converted to bytes before the
comparison and therefore always evaluated to true because
torch.cuda.mem_get_info(device)[0] returns a value expressed in bytes.
This patch does the conversion to bytes before the comparison.

Signed-off-by: Adam Miller <admiller@redhat.com>