Chunk docs on spaces for really long lines
While we still prefer chunking on paragraphs and newlines, also allow
chunking documents on spaces to handle cases where markdown files have
really long lines so that we don't blow up the context window.

This change adds a test for the same that asserts our chunked
documents are under an expected size. The test failed before adding
the chunking on spaces and passes afterwards.

Fixes #968

Signed-off-by: Ben Browning <bbrownin@redhat.com>