make ilab serve, chat, and generate support mixtral
both generate and serve already had a flag for the context size, meaning the user can override the default of 4096 in those two scenarios.

Added a new flag called `model-family` to serve which currently supports `merlinite` or `mixtral`. This is then used to load the proper chat template from a hardcoded map of supported templates. If nothing is specified, we default to the current one hardcoded in instructlab. any function that can start a server needs this flag as well. This meaans chat and generate need this flag.

Added the same flag to generate, since the generate prompt template needs extra padding for mixtral

Also, add `model-name` to `ilab train` as linux train had a hardcoded merlinite reference for the model name.

Most other "hardcoded" references are really just defaults which can be overridden by the user.

Signed-off-by: Charlie Doern <cdoern@redhat.com>