Introduce an end-to-end CI job on Linux
This PR introduces a new job that runs in CI that performs a minimal
configuration of the `ilab` workflow:

- `ilab init`
- `ilab download`
- `ilab serve`
- `ilab generate`
- `ilab train`

It runs on a GPU-enabled github actions runner. This is a Linux VM
with a single Tesla T4 GPU.

Given the resources required to run this job, we do not propose that
it runs automatically on every PR. Instead, it is a workflow that must
be manually launched. When launching it manually, you can specify
which branch or pull request to run it against.

Signed-off-by: Stef Walter <stefw@redhat.com>
Co-authored-by: Russell Bryant <rbryant@redhat.com>
Signed-off-by: Russell Bryant <rbryant@redhat.com>