Attempt to gracefully recover server context size errors
Sometimes, despite our best efforts at estimating maximum context size
and excluding large generated responses from including in the seed
examples, we still end up generating a prompt that exceeds the
server's context size. Instead of exiting the entire generation run,
attempt to gracefully log that it happened and continue generation.

There's no standard way servers communicate this, but in practice both
llama_cpp_python and vLLM use the same string in the error message to
indicate that this happened. So, that's what we key off of for now.

Signed-off-by: Ben Browning <bbrownin@redhat.com>