train rework, introduce --backend and --dtype flags
ilab train on macos and linux are really different just for MLX support.

It turns out pytorch supports an mps device enabling hardware accleration on macos. This allows us to use 99% of the codepath for linux train
with less storage and memory used. the mlx method requires adapters files, and a whole -fused dir that almost doubles the storage used in this whole process

This backend is also more maintainable as mlx is written in a way that lead us to need some heavy infrastructure to run it.

ilab train now defaults to --backend=pytorch but --backend=mlx is still available

added a few more enhancements:
- the --backend flag lets users choose pytorch or mlx
- I got rid of the multiple flags referencing "model" and just made a --model-repo flag since train should always pull the full safetensors from HF
- I got rid of a few bad flags people should probably never use like gguf-model-path gguf models for training yield horrible results most of the time
- --dtype takes fp16, bf16, ftp32, and auto. This sets the dtype to user in pytorch. bf16 and fp16 were hardcoded in the past.

Signed-off-by: Charlie Doern <cdoern@redhat.com>