feat: use labels to trigger backports
When setting a label to a PR, if it matches 'backport-release-v014'
Mergify will automatically create a backport PR in the configured stable
branch.
I've found this approach easier than writing a comment to a PR, also we
can visually see if a PR made it in a release branch by looking at the
label.
The only condition is to create the label.

Signed-off-by: Sébastien Han <seb@redhat.com>