docs: add inline mermaid diagram of basic workflow
Current workflow diagram is very detailed and complex.
The text above does not explain the items contained in the diagram.
The diagram is unclear for the beginner.
I understood the diagram only after reading all the documents.

The benefits of the proposed basic diagram:

It visualizes the text above.
It is simple for comprehension
It is easy to maintain
It fits better into the sequence of presentation.

Signed-off-by: Costa Shulyupin <costa.shul@redhat.com>