Only attempt clever model name determination for local `ilab serve`
This block of code switches the model name you see in `ilab chat` to
be what the server provides. We should only do this when we're talking
to a local `ilab serve` where we know the implementation doesn't care
what we send as the model name, and we know it is only serving up a
single model.

If we're not talking to our own local endpoint, we need to go with
what was provided on the command line or is in the chat config.

An improvement on top of this would be to do some validation on what
we're asking for against what's available in the remote endpoint.
There is a `404 Not Found` error we can predict ahead of time and give
a much nicer error message instead of just letting the api request
fail.

Signed-off-by: Russell Bryant <rbryant@redhat.com>