Fix up model name checking in the functional tests
This test was failing at the moment because it would spawn `ilab
serve` in the background and then immediately rename the model file. I
saw it failing because `ilab serve` was looking for the original
filename, but it tried to open it too late.

Switch the operations and test cases around a bit to be a bit more
deterministic and to match the current behavior of the code.

Co-authored-by: Russell Bryant <rbryant@redhat.com>
Signed-off-by: Sébastien Han <seb@redhat.com>