e2e: Disable pip cache
The caching from this job seems to be causing a problem with the rest
of the test jobs that do not install with cuda support. This is
causing CI failures where you see packages fail to load because NVIDIA
dependencies are not on the system.

The quick fix for the moment is to stop caching from this job. We'll
have to come back around and do a more focused cache specifically for
this.

Signed-off-by: Russell Bryant <rbryant@redhat.com>