ci: add 'one-approval' label on PR that has one approval
If a PR has been approved already by one of the maintainers, the PR will
be labeled with 'one-approval'. This is useful for maintainers to help
them determine which PR is getting close to completion and thus
prioritize the review.
With the help of filters, one can look for PRs that have one approval
already.

Signed-off-by: Sébastien Han <seb@redhat.com>