feat: ability to log server to file
Using --log-file, we can now write the server logs to a file. It's
going to be useful once we introduce support for systemd unit as well as
launchd. Also, having a log file is crucial step toward production
deployments and system integration.
Note that while --log-file is used, `ilab serve` will continue to log on
the foreground so if running inside a container you can get the
container log stream as well as a file.

Signed-off-by: Sébastien Han <seb@redhat.com>