README.md: Add a build note for Intel Macs
I just had a conversation on Slack where a user attempting to install
on an Intel Mac got a build error because it was trying to build with
Apple M-series GPU support. Adding an explicit build instruction to
disable that support made the build successful. Add this to the
install docs.

Signed-off-by: Russell Bryant <rbryant@redhat.com>