fix: do not fail when sending 'exit' to the chat
This is a partial revert of b8c9e63c550d07369f5381eb92b5285440cfa24b
which removes the unused 'content' argument of the `_handle_quit()`
function signature. Most of the time the function is not called with two
arguments unless the user types 'exit' in the chat, then `_handle_quit()`
is called with two arguments and we get:

```
TypeError: ConsoleChatBot._handle_quit() takes 1 positional argument but
2 were given```

To avoid linters complaining about an argument not being accessed we use
the  `_` keyword.

Signed-off-by: Sébastien Han <seb@redhat.com>