Explicitly set context size in length check test
Instead of assuming the default context size will always be 4096 or
smaller, explicitly set the server context size to 3072 both in the
server we start up when testing generation and in the `ilab generate`
parameter.

As a side-effect of this, the length of the question/answer pairs
required is slightly smaller than before. However, all Q/A pairs need
to be an appropriate length to trigger the bug this PR fixes, so the
overall length of the example long yaml file is larger because of
extending the size of the other Q/A pairs to be above this 3072 token
limit.

This also adjusts the test to match the new expected assert format,
which was needed because of rebasing this on top of all the changes in
the last month.

Signed-off-by: Ben Browning <bbrownin@redhat.com>