Add mypy typing checks to lint CI target
For starter, this patch lists all violating modules as excludes. The
list of excludes will be cleaned up over time as violations are
addressed.

The check is added to existing `lint` workflow since it doesn't
make the run much longer, and it's simpler than maintaining a separate
workflow that is largely the same. The new check will run regardless of
whether linters failures, as long as tox was successfully installed.

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>