sdg: Remove unnecessary config access
Calling `get_context()` with an arg of `DEFAULT_MULTIPROCESSING_START_METHOD`
is not necessary. The parent `lab.py` is already setting the default
before this code runs. Calling `get_context(None)` will return a
context using whatever the current default is set to.

This removes one bit of coupling from this sdg code and the rest of
the CLI, making it easier to extract into a library.

Signed-off-by: Russell Bryant <rbryant@redhat.com>