Fix ssl config when only cert file is passed
httpx.Client expects cert argument to be one of: a 2+ tuple, or a
string, or None.

Before the change, we were passing a 1-tuple when only the cert file was
configured through cli, which is violating the typing requirements of
the library; and it also made the library not initialize SSL context at
all. See:

https://github.com/encode/httpx/blob/37593c1952f4972040f6163da67e3777fd3d2e94/httpx/_config.py#L183
Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>