train: fix type violation when passing test_file and train_file
The linux_train function defines these arguments as `str`, but we are
now passing lists of strings. The `load_dataset` function down the
stream accepts either a `str` or a sequence of strings.

This patch changes the type of the arguments to `Path` since it's a
stronger type. It then converts to string before passing it further to
`load_dataset`.

This also reverts a test case change that made the check ineffective
because it no longer validated what the code under test did (instead
checking that pathlib library behavior is valid.)

This is a follow-up to e1d59105.

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>