ci: run e2e job when 2 approvals are given on a PR
The follow-up process occurs in two stages.  First, Mergify triggers the
end-to-end (e2e) workflow on a pull request (PR) after it receives two
approvals.  Once the e2e workflow runs on the PR, if it succeeds, the workflow
adds an 'e2e-success' label to the PR.  Mergify then has a new condition for
auto-merging, which relies on the presence of the 'e2e-success' label on the PR.

Signed-off-by: Sébastien Han <seb@redhat.com>