train: Reduce the memory check
On a test system with 64 GB RAM, this memory calculation came out as
62, not 64. Check for 60 instead of 64.

Obviously this is not very scientific as we're making very rough
assumptions about what is required. It would be better to enhance the
code further to actually calculate a memory requirement based on the
model instead just hard coding a rough guess.

Signed-off-by: Russell Bryant <rbryant@redhat.com>