Add aliasing to the old command structure
In order to alias properly, click requires a custom class that invokes a `get_command` function each time you call a click cmd. In my custom class, this function grabs a dictionary mapping cmd names to actual functions. These functions are then set to the actual cmd to be executed and returned. I also need to init the config here since
we technically skip over the "sub-parent" in this scenario. For example, in `ilab model list` the `model` part is where the config is checked into the context. We skip over that here, so it needs to happen in this get_command func.

Signed-off-by: Charlie Doern <cdoern@redhat.com>