create init_config
Previously, the config reading process and creation of the lab class could occur in lab.py. However, due to context passing norms in click, if we create the Lab object and
read the ocnfig into the context when creating the lab group, we lose the context after passing it down one level of command.
This means, we need to init our class object and context when at the sub-group level.

In this commit I moved this functionality and the Lab class into config.py so that we do not need to duplicate code in each submodule

Signed-off-by: Charlie Doern <cdoern@redhat.com>