mypy: Remove use of undefined GGUFWriter symbols
Neither add_expert_count nor add_expert_used_count are defined in the
version of gguf library we use.  Actually, these symbols are not yet
released on pypi, and are available in git repo of the library only:

https://github.com/ggerganov/llama.cpp/commit/799a1cb13b0b1b560ab0ceff485caed68faa8f1f

This makes mypy complain about it.

This patch removes code related to these parameters. If and when we
upgrade to gguf version that supports them, we may revert this patch.

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>