Use correct model family for granite
This change fixes a regression introduced in commit
bf332e84a3ddeaa4150bbeb6fb05648d6d0e8537 from PR #1174.

Prior to this commit, ilab would correctly use the `merlinte` model
family when using a granite model as that was the default. The change
in #1174 added some additional logic trying to guess the model family
based on filename to avoid having to specify `--model-family` in some
cases.

This change made the code automatically set the model family to
`granite`, which the code did not understand, which resulted in broken
model behavior.

The code has been updated to do the following:

- Understand that `granite` maps to `merlinite` for the model family

- When guessing a model family based on filename, don't assume the
  result is valid. Validate it against the set of known model families
  to `ilab` first.

- Fall back to the previous default of `merlinite` if the guess
  doesn't result in something `ilab` understands.

Fixes #1281

Signed-off-by: Russell Bryant <rbryant@redhat.com>
Signed-off-by: Guil Barros <gfbarros@bappos.com>