Handle connection exceptions to chat the model
When the server(endpoint) that serves the model is unreachable, we get
connection error. Until now we didn't handle this exception, causing a
full stack-trace, which can be hard to read and get to the real error.
Now, this exception is handled, giving the user the error and a hint of
the problem.

Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>