mergify: allow auto-merge for specific script without requiring lint/test checks
Updated Mergify configuration to support automatic merging when shell files
in `scripts/infra/` are modified. This change modifies the current
linter's regex to ignore all .sh files of not at the root of `scripts`.

This change aims to improve the merge workflow and reduce unnecessary
blocking for specific file changes.

Signed-off-by: Sébastien Han <seb@redhat.com>