local document source
Currently in the generation process: if any knowledge document or grounded skill is being processed: there is a hard dependency on needing to fetch the documents from the remote git source. For potential private taxonomy sources: this requires the ilab cli git processes to be configured to talk to the private git sources storing the documents. This introduces the ability to add a local document directory in the knowledge or grounded skill directory that can hold all document contents. The SDG process can then read all content from this directory meaning if the content has already been prepulled there is no need to dynmaically reach out to the remote git repo.

Signed-off-by: Tyler Lisowski <lisowski@us.ibm.com>