Add backend configuration and auto-detection support
This commit introduces the ability to configure and automatically detect
the backend for serving models in InstructLab. One new field, `backend`
has been added to the configuration and the CLI to allow users to
specify the backend. This field is optional.

Additionally, the commit includes the implementation of backend validation and
auto-detection based on the model file properties, specifically targeting GGUF
file format for now. The `serve` command has been updated to utilize these new
configurations and perform backend validation or auto-detection as required.

Unit tests for backend validation and changes to the configuration tests to
include the new fields are also part of this commit.

Co-authored-by: Ali Maredia <amaredia@redhat.com>
Signed-off-by: Sébastien Han <seb@redhat.com>