cli: Fix `--model-family granite`
Using a granite model works. Using a granite model plus specifying
`--model-family granite` did not work. This is because `granite` is
not one of the defined model families. Instead, we map `granite` to
`merlinite`, which in practice means an instructlab-tuned model.

Without explicitly specifying the model family optionn, we'd get
`granite` out of the model name and map it to the correct model
family. The code did not attempt to do this mapping when specified
explicitly.

This change makes the code check the model family mappings for
whatever was specified using `--model-family`, which will allow
`--model-family granite` to work.

Closes #1372

Signed-off-by: Russell Bryant <rbryant@redhat.com>