adapt backend auto-detection code to adopt vllm
The vLLM backend is being added to serve and chat with model. The
auto-detection code for which backend to use has been updated. Now we
will be able to detect which backend is best to be selected. vLLM runs
over a directory with tensor files in it and looks up for a config json
file. Our code only check that the model_path that was given points to a
directory, if it does we assume vLLM. We can definitly do better later.

Signed-off-by: Sébastien Han <seb@redhat.com>