give a generic name to backend logger
The logs from the backend server is being redirected from stdout/stderr
to an instance of a logger. When this happens the logger would print the
current log python file name. We replace this with "serving_backend" to
indicate that the log is coming from inside the uvicorn process that
runs the backend (llama_cpp or vllm).

Signed-off-by: Sébastien Han <seb@redhat.com>