disable 3.12 tests
instructlab/dolomite does not support Python 3.12 or 3.9. 3.9 could be removed fully as it is an older version. However, 3.12 support will eventually exist for dolomite.
linux train still works when using --legacy on 3.12 and macos train is fully in tact. For now, I am removing CI for 3.12 but keeping the docs and pyproject lines saying we support 3.12

Signed-off-by: Charlie Doern <cdoern@redhat.com>