adds e2e_data and e2e_checkpoint click options, validation
To support e2e training, we need a static place on the filesystem to
store intermediate checkpoints and our training data. The concept of a
checkpoints dir for other training paths is less strictly formed because
training is single-shot rather than phased in those regimes. To support
clean, expressive, user-friendly checkpoint caching (and ergonomic
serial evaluation) this crystalizes a location on the filesystem that
the e2e can expect is well-formed.

this also adds some degree of validation for the checkpoint and
data directories that the user specifies.

Signed-off-by: James Kunstle <jkunstle@redhat.com>